chore: bundle `valibot` (#4377)
<!-- Thank you for contributing! -->

### Description

Valibot is a library with quite a lot of surface area (2MB)

However, it is excellent for treeshaking and bundling - so this PR
bundles `valibot`.

This makes its size on disk for rolldown go from 2MB -> 33kB~

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated dependencies by removing unused packages from the browser
package and shifting certain dependencies in the rolldown package to
development-only status. No changes to user-facing functionality.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->