feat: expose `loadConfig` in api (#4428)
<!-- Thank you for contributing! -->

Closes #4313

### Description

This PR moves `loadConfig` from `cli/` to `utils/`, and exposes it to be
imported by users.

It also adds a `rolldown/config` entrypoint which would in theory allow
only importing code related to configuration rather than everything, but
I'm doubtful it's helpful when `loadConfig` imports the compiler
anyways.

Will test if it creates a smaller entrypoint, but if it doesn't or
maintainers also don't think it will be helpful I'll remove it

---------

Co-authored-by: Kevin Deng 三咲智子 <sxzz@sxzz.moe>