chore(rust): remove unused `tracing_unstable` cfg (#4510)
<!-- Thank you for contributing! -->

### Description

`RUSTFLAGS` would override the rustflags specified in
`.cargo/config.toml`, so I guess this should not be set.
https://doc.rust-lang.org/cargo/reference/config.html#buildrustflags


https://github.com/rolldown/rolldown/blob/513a17ae165b4ff12a3bebf169f159b3355651e9/.cargo/config.toml#L13-L57

refs https://github.com/rolldown/rolldown/issues/4508
This is the only place I found the difference between
`@rolldown/browser` and `@rolldown/binding-wasm32-wasi`.

---

`tracing_unstable` is for using `valuable` crate, which is removed at
https://github.com/rolldown/rolldown/pull/4253.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->

---------

Co-authored-by: Yunfei He <i.heyunfei@gmail.com>