revert: `feat(on_demand_wrapping): don't wrap modules that don't rely on others and have side effect #4670` (#4686)
<!-- Thank you for contributing! -->

### Description

Close https://github.com/rolldown/rolldown/issues/4684.

---

No side effect is not the full condition to safely remove the wrapper. A
module might not have side effects, but it might rely on the order that
side-effect module get executed first. See the case at
https://github.com/rolldown/rolldown/issues/4684.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->