feat: make native json plugin callable (#4908)
> By the way, are there any native plugins that can be used in dev other
than the resolver? If there aren't, I wonder if it's better to disable
them in `resolveConfig`.

Currently, we only have three native plugins that are potentially used
in the dev environment: `resolver`, `wasmFallback`, and maybe this
`json` plugin. The rest aren't callable in dev because they either use
the `transform_ast` hook, rely on internal context, or are only used in
build environment.

_Originally posted by @shulaoda in
https://github.com/vitejs/rolldown-vite/issues/246#issuecomment-2958188784_