perf: reduce temporary memory usage by avoiding output chunk clone until needed (#5058)
These two lines clones the whole bundle output. This is because
`args.bundle` is `Vec<Output>` where `Output` is a enum that contains a
`Box`.

https://github.com/rolldown/rolldown/blob/35d8e1569d820347646a014fca28ed32457f912c/crates/rolldown_binding/src/options/plugin/js_plugin.rs#L483

https://github.com/rolldown/rolldown/blob/35d8e1569d820347646a014fca28ed32457f912c/crates/rolldown_binding/src/options/plugin/js_plugin.rs#L510

While this `.clone` is inevitable to allow referencing the `bundle`
parameter of `generateBundle` hook after the hook has end, the content
of `bundle` parameter does not need to have a separate value, because it
is readonly (the modification is done via `ChangedOutputs` instead).
This PR changes `Box` to `Arc` so that we can share the underlying
value.

This PR will have a bigger impact if the output is large. For example,
`astro build` with https://github.com/withastro/astro.build used ~8GB of
memory previously (if the GC does not run frequently enough). With this
PR, it now uses ~3GB.

refs #1082