feat(rust): skip transform hook for modules starts with `rolldown:` (#5169)
Discussed at
https://github.com/rolldown/rolldown/pull/5159#issuecomment-3035306968.

On reflection, I think it's better to not allow it in the first place.
Considering possible `rolldown:oxc-runtime` in the future, it's not a
good idea to let users modify internal modules' code. @sapphi-red cc

However, if we do require this in the future, we could rethink this
decision.