chore(rollup-tests): skip occasionally failing test case (#5703)
This test case occasionally fails on CI, so it is temporarily skipped. 

I suspect the cause is a data race between synchronous and asynchronous
operations: on the NAPI side, `block_on` runs in NAPI’s async
environment, while the `emitFile` logic inside Rolldown runs in Tokio’s
async environment. This issue may be more likely to reproduce on
lower-end machines.