fix(hmr): ensure patch file name to be unique (#5825)
Patch file names were not ensured to be unique and caused flaky failures
in #5823.
The failure was caused by overriding a patch file with a newer one.

https://github.com/rolldown/rolldown/actions/runs/17120246519/job/48560035074#step:6:1526

The patch file name generation logic did not generate a unique name
because

- It uses
[`std::time::SystemTime`](https://doc.rust-lang.org/std/time/struct.SystemTime.html),
which is not monotonic.
- It uses millisecond precision.

This PR changes that logic to use a simple counter instead.