fix: trigger trace subscriber cleanup on Node.js side (#5913)
I was getting incomplete trace files when stopping the process by `Ctrl
+ C` or `process.exit()`.
I'm not sure the reason, but it seemed napi's env_cleanup_hook was not
called.
In this PR, I changed the code to trigger the trace subscriber cleanup
on Node.js side (`process.on('exit')` events using `signal-exit`).

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>