fix(dev): ensure patch file names to be unique (#6096)
The HMR manager is re-created for each update in the new dev API. This caused `next_hmr_patch_id` to be always `0`. That made the patch file to be always have the same name. This PR fixes that by sharing `next_hmr_patch_id` across all HMR managers.

Maybe this was making the tests flaky?

refs #5825