refactor: wait `this.load` using futures instead of callbacks (#6222)
The code for `this.load` waited for the `markModuleLoaded` callback to be called. But this should be able to simplify by using a future.
This PR refactors the code for `this.load` to wait for a future instead. This should also improve the perf a bit because we no longer have to call `markModuleLoaded` for each module.
I wonder if I'm missing something.

refs #4156, #2777, #2772, #2687