fix(dev): add `hasLatestBuildOutput` instead of `scheduleBuildIfStale` and add edit-reload test (#6321)
The following scenario was not working.

1. Edit file
2. HMR happens
3. Reload the window (stale content was shown!)

It is because the fix for `ensure_latest_build_output` in #6262 wasn't applied to `scheduleBuildIfStale`. To achieve that properly, it requires a significant refactor as `scheduleBuildIfStale` returns a future of the build process. For now, I added `hasLatestBuildOutput` instead to achieve a similar behavior. This approach has a problem of race condition because the internal state mutex is released in between the `hasLatestBuildOutput` call and the `ensureLatestBuildOutput` call. But this should be good enough for the initial implementation.