fix: allow specifying packageJsonPath in PluginContext.load (#6439)
For codes like below, the moduleDefFormat was not preserved.
```js
const resolved = await this.resolve('./dir/index.js', id)
await this.load(resolved)
```
This PR fixes that by receiving `package_json_path` in `this.load`.

If we receive `module_def_format`, it would make the code more simple. But I guess we don't want to expose that on JS side and keep it an internal thing.

refs #6398