docs: rework options documentation first round (#6542)
Preview:
https://deploy-preview-6542--rolldown-rs.netlify.app/options/input

The new documentation featured:

- Standalone/distributed pages for each option. This makes it friendly
for AI to consume
- Clean and unified structure
- The doc goes with simple -> in-depth content
- In-depth explanation to warn about edge cases or footguns.
- Glossary page for deep understanding (Also use distributed structure


## Todos

It's too much to gather all changes into one PR. Though I finished the
basics first, which is much better than the previous one.

Here are todos I record at
https://github.com/rolldown/rolldown/issues/6089#issuecomment-3418449667,
which will be solved in the subsquent PRs.

Missing page:

- transform
- plugins
- watch
- checks
- makeAbsoluteExternalsRelative
- debug

To be polished in the second rounds

- cwd
- platform
- shimMissingExports
- treeshake @IWANABETHATGUY 
- preserveEntrySignatures @IWANABETHATGUY 
- optimization.inlineConst @IWANABETHATGUY 
- logLevel
- onLog
- onwarn
- moduleTypes
- transform
- tsconfig
- all output options

To be improved (as a reminder)

- In `resolve options`'s in-depth, should warn about overwriten options
- For `extensionAlias`, should warn about adding `.js` and the trap of
resolving `.js` itself.
- For `extensionAlias`, the doc mentioned it could be used to handle ts.
Should invetigate about whether we already handle the situation.