chore: fix typo (#6801)
I was trying to open a PR from my vs code. But I was not able to open it
at all. I was pushing and publishing but there was no alert message or
no changes reflecting in the github like new branch, create a pr etc.
Then I need to change and open pr from the browser.

Signed-off-by: abeer0 <47961062+iiio2@users.noreply.github.com>