refactor(node): implement a new `Bundler` that satisfy the usage of `RolldownBuild` (#6877)
This's a transitive refactor and still some things need to be done in
later PRs. So

- watch mode is not affected because I isolated changes and only affect
the `RolldownBuild` logic.
- dev mode is not affected too.
- `Bundler` in `crates/rolldown` is not touched to avoid too much
changes.
- This PR doesn't seperate the bundler/build level data, but it's a
preparation for that purpose.


---

This PR introduces a `Build`​ concept to represent the process of doing
a build. Having this abstraction allows us to implement different
bundler for different usages.

In the previous code, `RolldownBuild`​ create a `Bundler`​ underlying
for each `generate/write`​ call. Without mentioning potential errors,
this abstraction is already wrong and give us buggy mental model when
working aounrd these code.

With this PR, `RolldownBuild`​ create only a `Bundler`​, but multiple
`builds`​ underlying for each `generate/write`​ call. This gives us a
correct mental model when working with code under `rolldown`​ without
needing to considering about the usage of `RolldownBuild`​.