refactor(rust): rename `BuildFactory/Build` to `BundleFactory/Bundle` (#6886)
I noticed this while I was working around the code. Rollup has the concept of `build`​ stage/phase. Rolldown followed this abstraction subconsciously.

The previous lifecycle of `Build`​ includes `build`​ + `generate`​ stages, which is improper, so we use `bundle`​ as the concept of finishing the process that starts from `build`​ and end with `generate`​.