feat(rust): use `BundleMode` to handle incremental build exhaustively (#6894)
Several reasons I want to introduce `BundleMode`:

## A mental model footgun

`BundleMode` is used to emphasize there are 3 states when we integrate incremental build into rolldown.

For exmaple, for watch mode:

1. if `incremental_build: false`, file changes should result to full rebuild every time
2. if `incremental_build: true`, the first build is not a plain full build, it needs to be done with saving data into cache.
3. if `incremental_build: true`, files changes should result to incremental rebuilds.

The footgun:

https://github.com/rolldown/rolldown/blob/c19666e8309871688d997729d32f49108604249a/crates/rolldown/src/watch/watcher_task.rs#L53-L67

When `incremental_build: false`, the watch mode will use `ScanMode:Full` in order to trigger full build, which seems to satisfy the above `1.`, but it's not!

> 1. if `incremental_build: false`, file changes should result to full rebuild every time

`watch mode` triggers a full build with extra unnecessary overhead of saving cache! And for each file change.

This PR doesn't fix the behavior but make it clear to be recognized. See https://github.com/rolldown/rolldown/pull/6894/files#r2507176596.

## Reduce bug prone patterns:

- https://github.com/rolldown/rolldown/pull/6894/files#r2507167806
- https://github.com/rolldown/rolldown/pull/6894/files#r2507174518

## Misc

Inside of the `Bundle` logic, we're using `ScanMode` + `is_incremental_build_enabled` to achieve the same effect as `BundleMode`. I restrict the refactor area of this PR, so they are not touched.

The previous design for the bundler hides many issues and cause advance features vulnerable. https://github.com/rolldown/rolldown/pull/6891#discussion_r2507078414 probably won't be the last issue.